+++ src/queries/payment/payment.model.ts
@@ -2,8 +2,10 @@ import { SortOrder } from '~/types/query.type'
 
 export const PAYMENTS_ENDPOINT = {
   default: '/payments',
-  order: (body: { order: SortOrder }) => {
-    return `${PAYMENTS_ENDPOINT.default}?order=date.${body.order}`
+  list: (year: number, month: number, sortOrder: SortOrder) => {
+    const startDate = `${year}-${month.toString().padStart(2, '0')}-01`
+    const endDate = `${year}-${month.toString().padStart(2, '0')}-${new Date(year, month, 0).getDate()}`
+    return `${PAYMENTS_ENDPOINT.default}?date=gte.${startDate}&date=lte.${endDate}&order=date.${sortOrder}`
   },
   detail: (paymentId: string) => {
     return `${PAYMENTS_ENDPOINT.default}?id=eq.${paymentId}&select=*,category:categories(*)&limit=1`
@@ -15,8 +17,8 @@ export const PAYMENTS_ENDPOINT = {
 
 export const PAYMENTS_KEY = {
   default: ['payment'],
-  order: (body: { order: SortOrder }) => {
-    return [...PAYMENTS_KEY.default, body.order]
+  list: (year: number, month: number, sortOrder: SortOrder) => {
+    return [...PAYMENTS_KEY.default, year, month, sortOrder]
   },
   detail: (paymentId: string) => {
     return [...PAYMENTS_KEY.default, paymentId]
